Added bulk edit for posts, created registration page, encrypting passwords with randSalt

+++ admin/includes/view_all_posts.php
@@ -1,6 +1,62 @@
+<?php
+
+if(isset($_POST['checkboxArray']))
+{
+    foreach($_POST['checkboxArray'] as $checkBoxValue)
+    {
+        $bulkOptions = $_POST['bulkOptions'];
+
+        switch ($bulkOptions) 
+        {
+            case 'Published':
+                $query = "UPDATE posts SET post_status = '{$bulkOptions}' WHERE post_id = {$checkBoxValue}";
+                $update_to_published = mysqli_query($connection, $query);
+                confirmQuery($update_to_published);
+                break;
+
+            case 'Draft':
+                $query = "UPDATE posts SET post_status = '{$bulkOptions}' WHERE post_id = {$checkBoxValue}";
+                $update_to_draft = mysqli_query($connection, $query);
+                confirmQuery($update_to_draft);
+                break;
+
+            case 'Delete':
+                $query = "DELETE FROM posts WHERE post_id = {$checkBoxValue}";
+                $delete_post = mysqli_query($connection, $query);
+                confirmQuery($delete_post);
+                break;
+
+            default:
+                # code...
+                break;
+        }
+    }
+}
+
+
+?>
+
+<h1 class="page-header">Posts</h1>
+
+<form action="" method="post">
 <table class="table table-bordered table-hover">
+
+    <div id="bulkOptionContainer" class="col-xs-4" style="padding-bottom: 15px;">
+        <select class="form-control" name="bulkOptions">
+            <option value="">Select Options</option>
+            <option value="Published">Publish</option>
+            <option value="Draft">Draft</option>
+            <option value="Delete">Delete</option>
+        </select>
+    </div>
+    <div class="col-xs-5">
+        <input type="submit" name="submit" class="btn btn-success" value="Apply">
+        <a class="btn btn-primary" href="posts.php?source=add_post">Add New</a>
+    </div>
+
     <thead>
         <tr>
+            <th><input id="selectAllBoxes" type="checkbox" name=""></th>
             <th>Post ID</th>
             <th>Author</th>
             <th>Title</th>
@@ -10,6 +66,7 @@
             <th>Tags</th>
             <th>Comments</th>
             <th>Date</th>
+            <th>View Post</th>
             <th>Edit</th>
             <th>Delete</th>
         </tr>
@@ -42,6 +99,12 @@
             }
 
             echo "<tr>";
+            ?>
+
+            <td><input class="checkBoxes" type="checkbox" name="checkboxArray[]" value="<?php echo $post_id; ?>"></td>
+
+
+            <?php
             echo "<td>{$post_id}</td>";
             echo "<td>{$post_author}</td>";
             echo "<td>{$post_title}</td>";
@@ -54,13 +117,15 @@
             echo "<td>{$post_tags}</td>";
             echo "<td>{$post_comment_count}</td>";
             echo "<td>{$post_date}</td>";
+            echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete the post(s)?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
             echo "</tr>";
         }
     ?>
     </tbody>
 </table>
+</form>
 
 <?php
